==> 05_Zero_shot_PII_masking_Using_LLM.ipynb <==
## This LLM(gpt-neo1.3B) hallucinates alot. I could not recieve Llama access on time, else I would have tried with Llama too.

from transformers import pipeline

# Initialize the GPT-2 pipeline
generator = pipeline("text-generation", model= "EleutherAI/gpt-neo-1.3B")


# Define the text you want to extract data from
input_text = """
You can also contact Alan Turing at alan.turing@example.org. 
"""
# Craft the prompt to instruct the model on what to do
prompt = f"""
Identify and mask all names and email addresses in the following text by replacing them with placeholders. Use [NAME] for names and [EMAIL] for email addresses. Ensure the output preserves the original structure and readability of the text.

Text: {input_text}
"""

# Use the model to generate the response
response = generator(prompt, max_length=200 , num_return_sequences=1)

# Print the structured response
print(response[0]['generated_text'])
